Extract Application logic from Client
